requirements.txt
asgiref==3.7.2
Django==4.2.5
django-jazzmin==2.6.0
Pillow==10.0.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
